fixed unsaved changes prompt
