+++ web/components/resources/Links.tsx
@@ -33,25 +33,6 @@ export default function Links(props: {
 
   useEffect(() => {
     props.setChangesMade(JSON.stringify(userChanges) !== JSON.stringify(links));
-
-    const warningText =
-      'You have unsaved changes - are you sure you wish to leave this page?';
-    const handleWindowClose = (e: BeforeUnloadEvent) => {
-      if (!props.getChangesMade()) return;
-      e.preventDefault();
-      return (e.returnValue = warningText);
-    };
-    const handleBrowseAway = () => {
-      if (!props.getChangesMade) return;
-      if (window.confirm(warningText)) return;
-      throw 'routeChange aborted.';
-    };
-    window.addEventListener('beforeunload', handleWindowClose);
-    router.events.on('routeChangeStart', handleBrowseAway);
-    return () => {
-      window.removeEventListener('beforeunload', handleWindowClose);
-      router.events.off('routeChangeStart', handleBrowseAway);
-    };
   }, [userChanges]);
 
   const updateLinks = async (newLinks: any) => {
