+++ web/components/resources/Research.tsx
@@ -1,11 +1,11 @@
-import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
-import ResearchURL from "./ResearchURL";
-import MdEditor from "react-markdown-editor-lite";
-import MarkdownIt from "markdown-it";
-import "react-markdown-editor-lite/lib/index.css";
-import { doc, getDoc, setDoc } from "firebase/firestore";
-import { db } from "@lib/firebase";
-import { useRouter } from "next/router";
+import React, { Dispatch, SetStateAction, useEffect, useState } from 'react';
+import ResearchURL from './ResearchURL';
+import MdEditor from 'react-markdown-editor-lite';
+import MarkdownIt from 'markdown-it';
+import 'react-markdown-editor-lite/lib/index.css';
+import { doc, getDoc, setDoc } from 'firebase/firestore';
+import { db } from '@lib/firebase';
+import { useRouter } from 'next/router';
 
 const mdParser = new MarkdownIt();
 
@@ -13,11 +13,11 @@ function Research(props: {
   getChangesMade: () => boolean;
   setChangesMade: Dispatch<SetStateAction<boolean>>;
 }) {
-  const [initialMarkdown, setInitialMarkdown] = useState("");
-  const [initialUrls, setInitialUrls] = useState([""]);
+  const [initialMarkdown, setInitialMarkdown] = useState('');
+  const [initialUrls, setInitialUrls] = useState(['']);
 
-  const [markdown, setMarkdown] = useState("");
-  const [urls, setUrls] = useState([""]);
+  const [markdown, setMarkdown] = useState('');
+  const [urls, setUrls] = useState(['']);
   const router = useRouter();
 
   useEffect(() => {
@@ -25,31 +25,12 @@ function Research(props: {
       JSON.stringify(urls) !== JSON.stringify(initialUrls) ||
         markdown !== initialMarkdown
     );
-
-    const warningText =
-      "You have unsaved changes - are you sure you wish to leave this page?";
-    const handleWindowClose = (e: BeforeUnloadEvent) => {
-      if (!props.getChangesMade()) return;
-      e.preventDefault();
-      return (e.returnValue = warningText);
-    };
-    const handleBrowseAway = () => {
-      if (!props.getChangesMade) return;
-      if (window.confirm(warningText)) return;
-      throw "routeChange aborted.";
-    };
-    window.addEventListener("beforeunload", handleWindowClose);
-    router.events.on("routeChangeStart", handleBrowseAway);
-    return () => {
-      window.removeEventListener("beforeunload", handleWindowClose);
-      router.events.off("routeChangeStart", handleBrowseAway);
-    };
   }, [markdown, urls, initialMarkdown, initialUrls]);
 
   useEffect(() => {
     let ignore = false;
 
-    getDoc(doc(db, "resources/research")).then((doc) => {
+    getDoc(doc(db, 'resources/research')).then((doc) => {
       if (!ignore) {
         setInitialMarkdown(doc?.data()?.markdown);
         setInitialUrls(doc?.data()?.url);
@@ -63,14 +44,13 @@ function Research(props: {
     };
   }, []);
   function setInfo() {
-    const researchDoc = doc(db, "resources", "research");
+    const researchDoc = doc(db, 'resources', 'research');
     setDoc(researchDoc, {
       markdown,
       url: urls,
     });
     setInitialMarkdown(markdown);
     setInitialUrls(urls);
-    console.log("reached here");
   }
   return (
     <div>
@@ -110,7 +90,7 @@ function Research(props: {
         <h2
           className="text-md mt-5 mb-5 font-bold text-blue-500 cursor-pointer pl-6"
           onClick={() => {
-            setUrls([...urls, ""]);
+            setUrls([...urls, '']);
           }}
         >
           + Add a link
@@ -120,8 +100,8 @@ function Research(props: {
             className={`py-2 px-3 rounded font-semibold hover:cursor-pointer border-[1px]
             ${
               props.getChangesMade()
-                ? "py-2 px-3 rounded border-[#304CD1] text-[#304CD1] hover:bg-[#304CD1] hover:text-[#ffffff] border-[1px] font-semibold hover:cursor-pointer"
-                : "py-2 px-3 rounded border-[#304CD1] text-[#304CD1] border-[1px] font-semibold hover:cursor-pointer"
+                ? 'py-2 px-3 rounded border-[#304CD1] text-[#304CD1] hover:bg-[#304CD1] hover:text-[#ffffff] border-[1px] font-semibold hover:cursor-pointer'
+                : 'py-2 px-3 rounded border-[#304CD1] text-[#304CD1] border-[1px] font-semibold hover:cursor-pointer'
             }`}
             onClick={setInfo}
           >
